netplay: v6 protocol — SessionTermination enum, narrowed Reject/Bye, M5 wire format
Coordinated v6 release covering M4 (reason codes) and M5 wire-format
foundation.  The wire is hard-broken from v5 (kProtocolVersion: 5→6
plus packet-size changes); mixed-version peers fail at decode rather
than via in-band negotiation.  All 50+ kReject*/kBye* call sites
keep compiling via constexpr aliases that resolve to the new enum
raw values.

M4 — reason codes (every callable site already used these):
- New `enum class SessionTermination : uint16_t` covering 28+ codes
  across handshake (1–9, preserves v5 numerics), snapshot (100–104),
  mid-session (200–203), transport (300–303), user/UI (400–402),
  broker UX (500–503), plus catchall.  Replaces the v5 `kReject*`
  and `kBye*` u32 enums; the historical kReject*/kBye* names live
  on as constexpr u16 aliases for source compatibility.
- NetReject.reason narrowed u32→u16 (struct now `{ u32 magic, u16 reason }`).
  kWireRejectSize: 8→6.
- NetBye gains a real u16 `reason` field (replacing the v5 unused
  4-byte tail).  kWireByeSize: 8→6.  v5 NetBye was magic-only; v6
  Bye now carries SessionTermination.
- Coordinator::SendReject / SendBye / End signatures narrowed
  uint32_t→uint16_t.  mLastRejectReason narrowed to match.  No
  call site changes — every emitter uses kReject*/kBye* aliases
  which are now u16.

M5 — observability wire format (active broadcast deferred):
- Three new packet types + magics:
    NetPhase ('ANPF', 12 B) — broadcast on every Coordinator phase
    transition + 1-Hz heartbeat in stable phases.
    NetEventBatch ('ANPV', 6-byte header + ≤28 × 8-byte events =
    ≤230 B) — batched fine-grained event stream.
    NetHeartbeat ('ANPT', 16 B) — 1-Hz RTT/loss/frame-skip telemetry.
- Encode/decode pairs in protocol.cpp + declarations in protocol.h.
  All three use the existing put_u16/get_u16 LE byte helpers; no
  new byte-IO infrastructure.
- protocol_selftest.cpp: testWireSizes asserts the new sizes; new
  testPhaseRoundTrip / testEventBatchRoundTrip / testHeartbeatRound-
  Trip mirror the testHelloRoundTrip pattern (round-trip + magic
  anchor + short-buffer rejection + bad-magic rejection).
  testRejectRoundTrip and testByeRoundTrip extended to cover the
  v6 narrowed reason field and the new SessionTermination values.
- Active broadcast hooks (SetPhaseAndBroadcast wrapping ~30
  mPhase = X transition sites, 1-Hz heartbeat scheduler in Poll,
  EmitEvent helper, peer-state caching, lobby SSE republish, page
  broker events consumer, in-emulator HUD) deferred — purely
  additive over the locked v6 wire and ships in a follow-up.

Verification:
- altirra-netplay-protocol-test: all round-trips pass with v6 sizes,
  including the cross-shape v5-vs-v6 sample.
- AltirraSDL native target: builds clean with the narrowed
  signatures (mechanical change preserved by the kReject*/kBye*
  aliases).
- altirra-lobby-test (45 tests) and altirra-lobby-ws-test (7
  tests): no regressions; wire change is invisible to lobby code.

v6 release coordination: bumping ALTIRRA_SDL_REF in altirra-sdl-
lobby's 4-pin (Dockerfile + 3 workflow YAMLs) ships the matching
lobby image; native Win/Linux/macOS/Android builds pick up the v6
constants from the same packets.h.  v5 peers refuse to handshake
with v6 peers (kRejectVersionSkew at HandleHelloFromJoiner) and
both sides cleanly fail at packet decode if the version skew is
ever bypassed.

+++ src/AltirraSDL/source/netplay/coordinator.cpp
@@ -2109,7 +2109,7 @@ void Coordinator::HandleBye(const NetBye& b, const Endpoint& from) {
 	}
 }
 
-void Coordinator::End(uint32_t byeReason) {
+void Coordinator::End(uint16_t byeReason) {
 	if (mPhase == Phase::Idle || mPhase == Phase::Ended || mPhase == Phase::Failed)
 		return;
 	// Tell every still-queued joiner that we're not coming — without
@@ -2142,7 +2142,7 @@ void Coordinator::End(uint32_t byeReason) {
 	}
 }
 
-void Coordinator::SendBye(uint32_t reason) {
+void Coordinator::SendBye(uint16_t reason) {
 	if (!mPeerKnown) return;
 	NetBye b { kMagicBye, reason };
 	size_t n = EncodeBye(b, mTxBuf, sizeof mTxBuf);
@@ -2236,7 +2236,7 @@ void Coordinator::SendWelcome(uint64_t nowMs) {
 	mWelcomeSentMs = nowMs;
 }
 
-void Coordinator::SendReject(uint32_t reason, const Endpoint& to) {
+void Coordinator::SendReject(uint16_t reason, const Endpoint& to) {
 	NetReject r { kMagicReject, reason };
 	size_t n = EncodeReject(r, mTxBuf, sizeof mTxBuf);
 	if (n) mTransport->SendTo(mTxBuf, n, to);
